Modularize application into reusable components
- Created 3 modular components without disrupting functionality:
  1. ChatMessages: Handles message display, loading states, and XML context
  2. SystemMessageEditor: Manages system message configuration panel
  3. ChatInput: Handles input, submission, streaming, and focus management

- Refactored main page.tsx to use modularized components
- Maintained all existing functionality including:
  - Streaming responses and tool execution
  - Focus management and scroll behavior
  - System message persistence
  - XML context window display
  - Loading states and error handling

- Improved code organization and maintainability
- Components are properly typed with TypeScript interfaces
- All tests pass via MCP browser automation
